requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.31.0
python-dotenv==1.0.0